change GetRoomRoles local
