--- app/definitions/IGetRoomRoles.ts
@@ -1,9 +1,0 @@
-export interface IGetRoomRoles {
-	_id: string;
-	rid: string;
-	u: {
-		_id: string;
-		username: string;
-	};
-	roles: string[];
-}
